# Run linters parallel