Before
---

```console
❯ hyperfine dist/lint
Benchmark 1: dist/lint
  Time (mean ± σ):     627.1 ms ±  23.5 ms    [User: 647.2 ms, System: 342.6 ms]
  Range (min … max):   596.3 ms … 667.8 ms    10 runs
```

After
---

```console
❯ hyperfine dist/lint
Benchmark 1: dist/lint
  Time (mean ± σ):     232.2 ms ±   7.6 ms    [User: 773.9 ms, System: 396.4 ms]
  Range (min … max):   217.3 ms … 245.6 ms    12 runs
```